# docs/openapi.yaml
openapi: 3.0.0
info:
  title: E-Commerce API
  version: 1.0.0
  description: REST API for an e-commerce application built with Node.js, Express, and PostgreSQL.
servers:
  - url: http://localhost:4001
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
...
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      responses:
        '200':
          description: List of products

    post:
      tags:
        - Products
      summary: Create a product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
                - inventoryQuantity
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                inventoryQuantity:
                  type: integer
      responses:
        '201':
          description: Product created successfully
        '400':
          description: Missing required fields
...
  /orders:
    get:
      tags:
        - Orders
      summary: Get all orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of orders
...